Add files via upload
Flood Fill is an algorithm typically used to find all the accessible areas from a starting point using depth-first search (DFS).

The idea is to give weight for each sell in the maze based on how far the sell from the goal, the goal cell have a weight of zero and the neighboring sells of the goal have a weight greater by one if those sells are not separated by wall, the neighbores of the ones will have a weight of two and so on.

When the robot is at the initial sell, it will move to the neighbor sell with least weight and continue like that until reaching the goal sell which have a zero weight.

The left hand rule is a also common strategy for solving mazes. It involves keeping your left hand against the wall of the maze and following it as you navigate through the maze. By always keeping your left hand on the wall, you will eventually reach the exit of the maze without getting lost. This strategy works because it ensures that you are always following a consistent path and will eventually cover the entire maze without retracing your steps.

The right hand rule is common strategy for solving mazes. It involves keeping your right hand against the wall of the maze and following it as you navigate through the maze. By always keeping your right hand on the wall, you will eventually reach the exit of the maze without getting lost. This strategy works because it ensures that you are always following a consistent path and will eventually cover the entire maze without retracing your steps.